Create an ObjectInspectorItem component (#7357)
* Rename and move the object inspector component into a components folder

* Create an ObjectInspectorItem component
