+++ packages/devtools-reps/src/object-inspector/components/ObjectInspector.css
